feat: update book shelf using BookAPI
Add a listener on Shelf that triggers a update on the Book associated
with it when the select values changes. Then we update the BookList
books state with the new shelf of the book.

Also there is a refactoring on the BookList that removes a unnecessary
logic related to the presenting the books by shelves.