refactor: attending to code review suggestions
on Shelf component: change the component to be a stateless one. Also fix
a bug on the select element options, that would not allow a book without
a shelf to be marked as "none" on his shelf.

on BookList component: change the component to be a stateless one.

on App component: handle the case where a book is marked on the "none"
shelf, which means the book is to be removed from all available shelves.

update the README.md to describe the project elements and how to install
and run it.